feat(tydoc,extractor): new packages "tydoc", "@tydoc/extractor"
The "tydoc" package is an entrypoint for most people. It has the CLI and
re-exports various Tydoc APIs that are re-exported from packages under
the @tydoc NPM namespace.

The "@tydoc/extractor" package is strictly an API (library) for getting the EDD.
