+++ packages/extractor/src/extractor/extract.ts
