+++ packages/extractor/src/index.ts
@@ -1,2 +1,2 @@
-export * from './api/extractor/extract'
-export { Options as RenderMarkdownOptions, render as renderMarkdown } from './api/renderers/markdown'
+export * from './extractor/extract'
+export { Options as RenderMarkdownOptions, render as renderMarkdown } from './renderers/markdown'
