+++ packages/tydoc/package.json
@@ -1,0 +1,61 @@
+{
+  "name": "tydoc",
+  "description": "The TypeScript documenter that meets you where you are",
+  "version": "0.3.0",
+  "main": "dist/index.js",
+  "license": "MIT",
+  "bin": "dist/cli/main.js",
+  "files": [
+    "dist",
+    "types.d.ts",
+    "types.js",
+    "oclif.manifest.json"
+  ],
+  "scripts": {
+    "build:oclif": "oclif-dev manifest",
+    "build:ts": "tsc",
+    "build": "yarn clean && yarn build:oclif && yarn build:ts",
+    "dev": "tsc --watch",
+    "clean": "rm -rf dist tsconfig.tsbuildinfo",
+    "dev:test": "jest --watch",
+    "test": "jest",
+    "test:dev": "yarn -s dev:test",
+    "format": "prettier --write .",
+    "prepublishOnly": "yarn build"
+  },
+  "dependencies": {
+    "@oclif/command": "^1.8.0",
+    "@oclif/config": "^1.17.0",
+    "@oclif/plugin-help": "^3.2.0",
+    "@tydoc/extractor": "workspace:packages/extractor",
+    "@types/debug": "^4.1.5",
+    "@types/lodash": "^4.14.165",
+    "debug": "^4.3.1",
+    "dedent": "^0.7.0",
+    "lodash": "^4.17.20"
+  },
+  "devDependencies": {
+    "@oclif/dev-cli": "1.24.1",
+    "@prisma-labs/prettier-config": "0.1.0",
+    "@types/dedent": "^0.7.0",
+    "@types/jest": "26.0.15",
+    "@types/node": "14.14.10",
+    "@types/prettier": "2.1.5",
+    "jest": "26.6.3",
+    "jest-validate": "26.6.2",
+    "jest-watch-suspend": "1.1.2",
+    "jest-watch-toggle-config": "2.0.1",
+    "jest-watch-typeahead": "0.6.1",
+    "ts-jest": "26.4.4",
+    "ts-node": "9.0.0",
+    "typescript": "^4.1.2"
+  },
+  "prettier": "@prisma-labs/prettier-config",
+  "oclif": {
+    "commands": "./dist/cli/commands",
+    "bin": "tydoc",
+    "plugins": [
+      "@oclif/plugin-help"
+    ]
+  }
+}
