+++ packages/tydoc/source/index.ts
@@ -1,0 +1,1 @@
+export * from '@tydoc/extractor'
