+++ packages/tydoc/source/types.ts
@@ -1,0 +1,1 @@
+export * as Doc from '@tydoc/extractor/types'
