+++ packages/web/CONTRIBUTING.md
@@ -1,0 +1,11 @@
+# Contributing Guide for TyDoc Web App
+
+### Overview
+
+- Next.js based app
+
+### Quickstart
+
+```
+yarn dev
+```
