+++ packages/web/components/Type.tsx
@@ -1,7 +1,6 @@
+import { Doc } from '@tydoc/extractor/types'
 import React, { FC } from 'react'
-import { Doc } from 'tydoc/types'
 import { CodeBlock } from './CodeBlock'
-
 import { Node } from './Node'
 
 interface TypeProps {
