+++ packages/web/pages/npm/[entrypoint]/[...pkg].tsx
@@ -1,8 +1,8 @@
+import { Layout } from '.../../../components/Layout'
+import { Doc } from '@tydoc/extractor/types'
 import { GetStaticPaths, InferGetStaticPropsType } from 'next'
 import React, { FC } from 'react'
-import { Doc } from 'tydoc/types'
 import { Package } from '../../../components/Package'
-import { Layout } from '.../../../components/Layout'
 import { defineStaticProps } from '../../../utils/next'
 import { NPM } from '../../../utils/types'
 import { getJson } from '../../../utils/utils'
